Refactor reservation and menu handling
- Changed `AdultsQty` and `KidsQty` data types in `Reservation.cs`
  from `decimal` to `long` for better alignment with expected
  integer values.
- Updated `MenuCategories` in `AddMenu.xaml` and `EditMenu.xaml`
  to simplify and reorganize the list of categories.
- Removed the "Checklist PDF" button from `ReservationDetails.xaml`
  as the feature is no longer required.
- Added logic in `ReservationDetailsViewModel.cs` to update the
  reservation status to "contractsigning" after sending the contract
  email.
- Improved logging in `ReservationDetailsViewModel.cs` to include
  status updates for better traceability.

+++ Models/Reservation.cs
@@ -32,10 +32,10 @@ namespace CATERINGMANAGEMENT.Models
         public string Location { get; set; } = string.Empty;
 
         [Column("adults_qty")]
-        public decimal AdultsQty { get; set; }
+        public long AdultsQty { get; set; }
 
         [Column("kids_qty")]
-        public decimal KidsQty { get; set; }
+        public long KidsQty { get; set; }
 
         [Column("status")]
         public string Status { get; set; } = string.Empty;
